FP6 second try

+++ yatube/posts/views.py
@@ -1,11 +1,13 @@
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, redirect, render
+from django.views.decorators.cache import cache_page
 
 from .forms import PostForm, CommentForm
 from .models import Group, Post, User, Follow
 from .utils import page_obj
 
 
+@cache_page(20)
 def index(request):
     context = {'page_obj': page_obj(request, Post.objects.all())}
     return render(request, 'posts/index.html', context)
@@ -100,5 +102,6 @@ def profile_unfollow(request, username):
         user=request.user,
         author__username=username
     )
-    user_follower.delete()
+    if user_follower.DoesNotExist():
+        user_follower.delete()
     return redirect('posts:profile', username)
